requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements/requirements.in
#
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.2.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.10.0
    # via -r requirements/requirements.in
certifi==2021.10.8
    # via urllib3
cffi==1.15.0
    # via
    #   cryptography
    #   trio
cryptography==35.0.0
    # via
    #   pyopenssl
    #   urllib3
h11==0.12.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
mysql-connector-python==8.0.27
    # via -r requirements/requirements.in
outcome==1.1.0
    # via trio
protobuf==3.19.1
    # via
    #   -r requirements/requirements.in
    #   mysql-connector-python
pycparser==2.21
    # via cffi
pyopenssl==21.0.0
    # via urllib3
selenium==4.0.0
    # via -r requirements/requirements.in
six==1.16.0
    # via pyopenssl
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.3
    # via beautifulsoup4
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure]==1.26.7
    # via selenium
wsproto==1.0.0
    # via trio-websocket
